refactor(MessageInput): move logic to hook
